feat: add create pages for both items
