+++ lib/ui/views/supplies/new_generic_item_specifics_page.dart
@@ -1,0 +1,101 @@
+import 'package:flutter/material.dart';
+import 'package:mona/data/model/generic_supply_item.dart';
+import 'package:mona/data/providers/supply_item_provider.dart';
+import 'package:mona/l10n/build_context_extensions.dart';
+import 'package:mona/ui/widgets/forms/form_dropdown_field.dart';
+import 'package:mona/ui/widgets/forms/form_text_field.dart';
+import 'package:mona/ui/widgets/forms/model_form.dart';
+import 'package:mona/util/string_parsing.dart';
+import 'package:provider/provider.dart';
+
+class NewGenericItemSpecificsPage extends StatefulWidget {
+  final String name;
+
+  const NewGenericItemSpecificsPage({super.key, required this.name});
+
+  @override
+  State<NewGenericItemSpecificsPage> createState() =>
+      _NewGenericItemSpecificsPageState();
+}
+
+class _NewGenericItemSpecificsPageState
+    extends State<NewGenericItemSpecificsPage> {
+  late TextEditingController _amountController;
+  GenericSupplyType? _genericSupplyType;
+
+  String? get _amountError =>
+      GenericSupply.validateAmount(context.l10n, _amountController.text);
+  String? get _genericSupplyTypeError =>
+      _genericSupplyType == null ? context.l10n.requiredField : null;
+
+  bool get _isFormValid =>
+      _amountError == null && _genericSupplyTypeError == null;
+
+  void _refresh() => setState(() {});
+
+  void _closeAll() {
+    Navigator.of(context)
+      ..pop()
+      ..pop();
+  }
+
+  void _addItem() {
+    final item = GenericSupply(
+      name: widget.name,
+      amount: _amountController.text.toInt,
+      genericSupplyType: _genericSupplyType!,
+    );
+    Provider.of<SupplyItemProvider>(context, listen: false).add(item);
+
+    Navigator.of(context)
+      ..pop()
+      ..pop();
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _amountController = TextEditingController();
+  }
+
+  @override
+  void dispose() {
+    _amountController.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final localizations = context.l10n;
+
+    return ModelForm(
+      title: widget.name,
+      submitButtonLabel: localizations.add,
+      isFormValid: _isFormValid,
+      saveChanges: _addItem,
+      closeAll: _closeAll,
+      fields: [
+        FormDropdownField<GenericSupplyType>(
+          value: _genericSupplyType,
+          items: GenericSupplyType.values
+              .map(
+                (type) => DropdownMenuItem<GenericSupplyType>(
+                  value: type,
+                  child: Text(type.name),
+                ),
+              )
+              .toList(),
+          onChanged: (value) => setState(() => _genericSupplyType = value),
+          label: 'Type',
+        ),
+        FormTextField(
+          controller: _amountController,
+          label: localizations.amount,
+          onChanged: _refresh,
+          inputType: TextInputType.number,
+          regexFormatter: r'[0-9]',
+        ),
+      ],
+    );
+  }
+}
