+++ lib/ui/views/supplies/new_item_page.dart
@@ -1,22 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:mona/data/model/administration_route.dart';
-import 'package:mona/data/model/ester.dart';
-import 'package:mona/data/model/medication_supply_item.dart';
-import 'package:mona/data/model/molecule.dart';
+import 'package:m3e_core/m3e_core.dart';
 import 'package:mona/data/model/supply_item.dart';
-import 'package:mona/data/providers/supply_item_provider.dart';
 import 'package:mona/l10n/build_context_extensions.dart';
-import 'package:mona/l10n/helpers/administration_route_l10n.dart';
-import 'package:mona/services/preferences_service.dart';
-import 'package:mona/ui/widgets/dropdowns/administration_route_dropdown.dart';
-import 'package:mona/ui/widgets/dropdowns/ester_dropdown.dart';
-import 'package:mona/ui/widgets/dropdowns/molecule_dropdown.dart';
-import 'package:mona/ui/widgets/forms/form_dropdown_field.dart';
+import 'package:mona/ui/views/supplies/new_generic_item_specifics_page.dart';
+import 'package:mona/ui/views/supplies/new_medication_item_specifics_page.dart';
 import 'package:mona/ui/widgets/forms/form_spacer.dart';
 import 'package:mona/ui/widgets/forms/form_text_field.dart';
 import 'package:mona/ui/widgets/forms/model_form.dart';
-import 'package:mona/util/string_parsing.dart';
-import 'package:provider/provider.dart';
+
+enum _ItemType { medication, genericSupply }
 
 class NewItemPage extends StatefulWidget {
   @override
@@ -24,115 +16,39 @@ class NewItemPage extends StatefulWidget {
 }
 
 class _NewItemPageState extends State<NewItemPage> {
-  late TextEditingController _totalAmountController;
   late TextEditingController _nameController;
-  late TextEditingController _concentrationController;
-  Molecule? _molecule;
-  AdministrationRoute? _administrationRoute;
-  Ester? _ester;
-  late PreferencesService _preferencesService;
+  _ItemType _type = _ItemType.medication;
 
   String? get _nameError =>
       SupplyItem.validateName(context.l10n, _nameController.text);
-  String? get _totalAmountError => MedicationSupplyItem.validateTotalAmount(
-      context.l10n, _totalAmountController.text);
-  String? get _concentrationError => MedicationSupplyItem.validateConcentration(
-      context.l10n, _concentrationController.text);
-  String? get _moleculeError =>
-      MedicationSupplyItem.validateMolecule(context.l10n, _molecule);
-  String? get _administrationRouteError =>
-      MedicationSupplyItem.validateAdministrationRoute(
-          context.l10n, _administrationRoute);
-  String? get _esterError {
-    final validator = MedicationSupplyItem.esterValidator(
-        context.l10n, _molecule, _administrationRoute);
-    return validator(_ester);
-  }
 
-  bool get _isFormValid =>
-      _nameError == null &&
-      _totalAmountError == null &&
-      _concentrationError == null &&
-      _moleculeError == null &&
-      _administrationRouteError == null &&
-      _esterError == null;
+  bool get _isFormValid => _nameError == null;
 
-  bool get _useEsterField =>
-      _molecule == KnownMolecules.estradiol &&
-      _administrationRoute == AdministrationRoute.injection;
+  void _refresh() => setState(() {});
 
-  void _onMoleculeChanged(Molecule? molecule) {
-    if (molecule != null) {
-      setState(() {
-        _molecule = molecule;
-
-        if (!_useEsterField) {
-          _ester = null;
-        }
-      });
-    }
-  }
-
-  void _onAdministrationRouteChanged(AdministrationRoute? administrationRoute) {
-    if (administrationRoute != null) {
-      setState(() {
-        _administrationRoute = administrationRoute;
-
-        if (!_useEsterField) {
-          _ester = null;
-        }
-      });
-    }
-  }
-
-  void _onEsterChanged(Ester? ester) {
-    if (ester != null) {
-      setState(() {
-        _ester = ester;
-      });
-    }
-  }
-
-  void _refresh() {
-    setState(() {});
-  }
-
-  void _addItem() async {
-    final totalAmount = _totalAmountController.text.toDecimal;
-    final concentration = _concentrationController.text.toDecimal;
-    final totalDose = concentration * totalAmount;
+  void _next() {
     final name = _nameController.text;
-    final supplyItemProvider =
-        Provider.of<SupplyItemProvider>(context, listen: false);
 
-    final item = MedicationSupplyItem(
-      name: name,
-      totalDose: totalDose,
-      concentration: concentration,
-      molecule: _molecule!,
-      administrationRoute: _administrationRoute!,
-      ester: _ester,
-    );
-    supplyItemProvider.add(item);
-
-    Navigator.pop(context);
+    final route = switch (_type) {
+      _ItemType.medication => MaterialPageRoute<void>(
+          builder: (context) => NewMedicationItemSpecificsPage(name: name),
+        ),
+      _ItemType.genericSupply => MaterialPageRoute<void>(
+          builder: (context) => NewGenericItemSpecificsPage(name: name),
+        ),
+    };
+    Navigator.of(context).push(route);
   }
 
   @override
   void initState() {
     super.initState();
-    _preferencesService =
-        Provider.of<PreferencesService>(context, listen: false);
-    _totalAmountController = TextEditingController();
     _nameController = TextEditingController();
-    _concentrationController = TextEditingController();
   }
 
   @override
   void dispose() {
-    _totalAmountController.dispose();
     _nameController.dispose();
-    _concentrationController.dispose();
     super.dispose();
   }
 
@@ -142,9 +58,9 @@ class _NewItemPageState extends State<NewItemPage> {
 
     return ModelForm(
       title: localizations.newItem,
-      submitButtonLabel: localizations.add,
+      submitButtonLabel: localizations.next,
       isFormValid: _isFormValid,
-      saveChanges: _addItem,
+      saveChanges: _next,
       fields: [
         FormTextField(
           controller: _nameController,
@@ -153,47 +69,25 @@ class _NewItemPageState extends State<NewItemPage> {
           inputType: TextInputType.text,
         ),
         FormSpacer(),
-        FormDropdownField<Molecule>(
-          value: _molecule,
-          items: moleculeDropdownMenuItems(
-            _preferencesService.allMolecules,
-            localizations,
-          ),
-          onChanged: _onMoleculeChanged,
-          label: localizations.molecule,
-        ),
-        FormDropdownField<AdministrationRoute>(
-          value: _administrationRoute,
-          items: administrationRouteDropdownMenuItems(localizations),
-          onChanged: _onAdministrationRouteChanged,
-          label: localizations.adminRoute,
-        ),
-        if (_useEsterField)
-          FormDropdownField<Ester>(
-            value: _ester,
-            items: esterDropdownMenuItems(localizations),
-            onChanged: _onEsterChanged,
-            label: localizations.ester,
-          ),
-        FormSpacer(),
-        FormTextField(
-          controller: _totalAmountController,
-          label: localizations.totalAmount,
-          onChanged: _refresh,
-          inputType: TextInputType.numberWithOptions(decimal: true),
-          suffixText: _administrationRoute?.localizedUnit(localizations, 1),
-          regexFormatter: r'[0-9.,]',
-        ),
-        FormTextField(
-          controller: _concentrationController,
-          label: localizations.concentration,
-          onChanged: _refresh,
-          inputType: TextInputType.numberWithOptions(decimal: true),
-          suffixText: _molecule != null && _administrationRoute != null
-              ? '${_molecule!.unit}/${_administrationRoute!.localizedUnit(localizations, 1)}'
-              : null,
-          regexFormatter: r'[0-9.,]',
-        ),
+        _typeToggle(),
+      ],
+    );
+  }
+
+  Widget _typeToggle() {
+    return M3EToggleButtonGroup(
+      type: M3EButtonGroupType.standard,
+      size: M3EButtonSize.md,
+      selectedIndex: _type.index,
+      onSelectedIndexChanged: (index) {
+        if (index == null) return;
+        setState(() {
+          _type = _ItemType.values[index];
+        });
+      },
+      actions: const [
+        M3EToggleButtonGroupAction(label: Text('Medication')),
+        M3EToggleButtonGroupAction(label: Text('Generic supply')),
       ],
     );
   }
